move indexing function into its own Python module
This makes it mow a standard function of our new Python
library instead of a stand-alone program.
